My fellow citizens, at this hour, American and coalition forces are in the early stages of military operations to disarm Iraq, to free its people and to defend the world from grave danger.
On my orders, coalition forces have begun striking selected targets of military importance to undermine Saddam Hussein's ability to wage war. These are opening stages of what will be a broad and concerted campaign. More than 35 countries are giving crucial support -- from the use of naval and air bases, to help with intelligence and logistics, to the deployment of combat units. Every nation in this coalition has chosen to bear the duty and share the honor of serving in our common defense.

To all the men and women of the United States Armed Forces now in the Middle East, the peace of a troubled world and the hopes of an oppressed people now depend on you. That trust is well placed.

The enemies you confront will come to know your skill and bravery. The people you liberate will witness the honorable and decent spirit of the American military. In this conflict, America faces an enemy who has no regard for conventions of war or rules of morality. Saddam Hussein has placed Iraqi troops and equipment in civilian areas, attempting to use innocent men, women and children as shields for his own military -- a final atrocity against his people.